requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
litellm>=1.40.0
streamlit>=1.35.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2